Fix entities

+++ src/main/java/org/princelle/lp1project/Entities/AchievedService.java
@@ -11,7 +11,7 @@ public class AchievedService {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private long id;
 
 	private Date date;
 
@@ -20,13 +20,9 @@ public class AchievedService {
 	private boolean validTo;
 	private boolean validFrom;
 
-	public int getId() {
-		return id;
-	}
+	public long getId() { return id; }
 
-	public void setId(int id) {
-		this.id = id;
-	}
+	public void setId(int id) { this.id = id; }
 
 	public Date getDate() {
 		return this.date;
