+++ src/main/java/org/princelle/lp1project/Entities/Colocation.java
@@ -18,13 +18,9 @@ public class Colocation {
 	@JoinColumn(name = "member")
 	private List member;
 
-	public long getId() {
-		return id;
-	}
+	public long getId() { return id; }
 
-	public void setId(long id) {
-		this.id = id;
-	}
+	public void setId(int id) { this.id = id; }
 
 	public List<Person> getPeople() {
 		return member;
