+++ src/main/java/org/princelle/lp1project/Entities/Person.java
@@ -25,13 +25,9 @@ public class Person {
 	@Column(name = "score", nullable = false)
 	private Integer score = 0;
 
-	public long getId() {
-		return id;
-	}
+	public long getId() { return id; }
 
-	public void setId(long id) {
-		this.id = id;
-	}
+	public void setId(int id) { this.id = id; }
 
 	public Integer getScore() {
 		return score;
